feat: :sparkles: manejo de errores

+++ src/main/java/com/ponscio/model/Empleado.java
@@ -1,12 +1,13 @@
 package com.ponscio.model;
+import java.math.BigDecimal;
 import java.util.Map;
 
 public class Empleado {
     private int id, rol_id;
     private String nombre, correo, documento_numero, documento_tipo;
-    private double salario;
+    private BigDecimal salario;
     
-    public Empleado(int id, String nombre, String documento_numero, String documento_tipo,int rol_id, String correo, double salario) {
+    public Empleado(int id, String nombre, String documento_numero, String documento_tipo,int rol_id, String correo, BigDecimal salario) {
         this.id = id;
         this.nombre = nombre;
         this.documento_numero = documento_numero;
@@ -34,7 +35,7 @@ public class Empleado {
     public String getCorreo() {
         return correo;
     }
-    public double getSalario() {
+    public BigDecimal getSalario() {
         return salario;
     }
     
@@ -47,7 +48,7 @@ public class Empleado {
         info += "\n\tDocumento: "+documento_numero;
         info += "\n\tCorreo: "+correo;
         
-        info += "\n\tSalario actual: "+salario;
+        info += "\n\tSalario actual: "+ salario;
         info += "\n\tCargo Actual: "+ roles.get(rol_id);
         info += "\n\n==========================================";
         return info;
